Get rid of 2 pass razor pages logic in the default case - Today RazorPages uses a 2 pass model where initially pages are discovered but not "compiled" up front. Then a matcher policy is used to build the compiled page action descriptor with the type information. This requires running logic per request that determins if we need to compile the page. By default, runtime compilation is off and the compiled page information is available via a dictionary lookup. We can use the page loader at startup time to avoid doing lazy lookup of this information and a dynamic swap in the endpoint system. Instead we can build the endpoint for the compiled page descriptor directly.
